+++ src/Admin/Tags/Events/TagGroupCreated.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Thinktomorrow\Chief\Admin\Tags\Events;
+
+use Thinktomorrow\Chief\Admin\Tags\TagGroupId;
+
+class TagGroupCreated
+{
+    public function __construct(public readonly TagGroupId $tagGroupId)
+    {
+
+    }
+}