+++ src/Admin/Tags/Repositories/MediaTagRepository.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Thinktomorrow\Chief\Admin\Tags\Repositories;
-
-class MediaTagRepository
-{
-}