+++ src/Admin/Tags/Taggable.php
@@ -2,9 +2,14 @@
 
 namespace Thinktomorrow\Chief\Admin\Tags;
 
+use Illuminate\Database\Eloquent\Relations\MorphToMany;
+use Illuminate\Support\Collection;
+use Thinktomorrow\Chief\Admin\Tags\Read\TagRead;
+
 interface Taggable
 {
-    public function getTaggableType(): string;
+    /** @return TagRead[] */
+    public function getTags(): Collection;
 
-    public function getTaggableId(): string;
+    public function tags(): MorphToMany;
 }