+++ tests/Application/Admin/Tags/TaggableTest.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Thinktomorrow\Chief\Tests\Application\Admin\Tags;
+
+use Thinktomorrow\Chief\Tests\Application\Admin\Tags\Crud\TagTestHelpers;
+use Thinktomorrow\Chief\Tests\ChiefTestCase;
+
+class TaggableTest extends ChiefTestCase
+{
+    use TagTestHelpers;
+
+    public function test_taggable_can_get_all_tags()
+    {
+
+    }
+}