+++ tests/Unit/Admin/Tags/TagTest.php
@@ -3,18 +3,19 @@
 namespace Thinktomorrow\Chief\Tests\Unit\Admin\Tags;
 
 use Thinktomorrow\Chief\Admin\Tags\Tag;
-use Thinktomorrow\Chief\Admin\Tags\TagState;
+use Thinktomorrow\Chief\Admin\Tags\TagGroupId;
+use Thinktomorrow\Chief\Admin\Tags\TagId;
 use Thinktomorrow\Chief\Tests\TestCase;
 
 class TagTest extends TestCase
 {
     public function test_it_can_create_a_tag()
     {
-        $tag = new Tag('id', TagState::online, 'intern label', ['option' => 'foobar']);
+        $tag = new Tag($tagId = TagId::fromString('xxx'), $tagGroupId = TagGroupId::fromString('yyy'), 'intern label', ['option' => 'foobar']);
 
 
-        $this->assertEquals('id', $tag->id);
-        $this->assertEquals(TagState::online, $tag->getState());
+        $this->assertEquals($tagId, $tag->tagId);
+        $this->assertEquals($tagGroupId, $tag->getTagGroupId());
         $this->assertEquals('intern label', $tag->getLabel());
         $this->assertEquals(['option' => 'foobar'], $tag->getData());
     }