+++ CHANGELOG.md
@@ -14,6 +14,7 @@ principles.
 -   Fixed: field-in-window flexbox items are sometimes overflowing the container (e.g. repeat field window view)
 -   Fixed: RadioFilter and CheckboxFilter field labels were not correctly styled
 -   Fixed: Checkbox form-input-toggle doesn't shrink anymore in flexbox with long label
+- Fixed: charactercount was not triggered in forms on initial page DOM, only in sidebar.
 
 ## 0.7.22 - 2023-03-13